+++ molgenis-app/src/main/java/org/molgenis/app/WebAppPermissionRegistry.java
@@ -8,7 +8,7 @@ import org.molgenis.data.DataService;
 import org.molgenis.data.plugin.model.PluginIdentity;
 import org.molgenis.data.security.auth.Role;
 import org.molgenis.data.security.auth.User;
-import org.molgenis.security.acl.SidUtils;
+import org.molgenis.data.security.SidUtils;
 import org.molgenis.security.core.PermissionSet;
 import org.molgenis.util.Pair;
 import org.springframework.security.acls.model.ObjectIdentity;
@@ -21,7 +21,7 @@ import static org.molgenis.data.security.auth.RoleMetadata.ROLE;
 import static org.molgenis.data.security.auth.UserMetaData.USER;
 import static org.molgenis.data.security.auth.UserMetaData.USERNAME;
 import static org.molgenis.security.account.AccountService.ROLE_USER;
-import static org.molgenis.security.acl.SidUtils.createRoleSid;
+import static org.molgenis.data.security.SidUtils.createRoleSid;
 import static org.molgenis.security.core.utils.SecurityUtils.ANONYMOUS_USERNAME;
 
 /**
