+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/DataServiceIT.java
@@ -17,8 +17,8 @@ import org.molgenis.data.support.AggregateQueryImpl;
 import org.molgenis.data.support.QueryImpl;
 import org.molgenis.data.util.EntityUtils;
 import org.molgenis.data.validation.MolgenisValidationException;
-import org.molgenis.security.PermissionService;
-import org.molgenis.security.acl.SidUtils;
+import org.molgenis.security.core.PermissionService;
+import org.molgenis.data.security.SidUtils;
 import org.molgenis.security.core.PermissionSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.acls.model.ObjectIdentity;
