+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/MetaDataServiceIT.java
@@ -16,8 +16,8 @@ import org.molgenis.data.support.DynamicEntity;
 import org.molgenis.data.support.EntityWithComputedAttributes;
 import org.molgenis.data.util.EntityUtils;
 import org.molgenis.data.util.MolgenisDateFormat;
-import org.molgenis.security.PermissionService;
-import org.molgenis.security.acl.SidUtils;
+import org.molgenis.security.core.PermissionService;
+import org.molgenis.data.security.SidUtils;
 import org.molgenis.security.core.PermissionSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.acls.model.ObjectIdentity;
