+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/importservice/EmxImportServiceIT.java
@@ -14,8 +14,8 @@ import org.molgenis.data.meta.model.Tag;
 import org.molgenis.data.security.EntityTypeIdentity;
 import org.molgenis.data.security.PackageIdentity;
 import org.molgenis.data.security.auth.User;
-import org.molgenis.security.PermissionService;
-import org.molgenis.security.acl.SidUtils;
+import org.molgenis.security.core.PermissionService;
+import org.molgenis.data.security.SidUtils;
 import org.molgenis.security.core.PermissionSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.acls.model.ObjectIdentity;
