+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/importservice/OntologyImportServiceIT.java
@@ -8,7 +8,7 @@ import org.molgenis.data.importer.ImportService;
 import org.molgenis.data.security.EntityTypeIdentity;
 import org.molgenis.data.security.auth.User;
 import org.molgenis.ontology.core.meta.Ontology;
-import org.molgenis.security.PermissionService;
+import org.molgenis.security.core.PermissionService;
 import org.molgenis.security.core.PermissionSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.acls.model.ObjectIdentity;
@@ -23,7 +23,7 @@ import static java.util.Collections.emptySet;
 import static java.util.Objects.requireNonNull;
 import static org.molgenis.data.DatabaseAction.ADD;
 import static org.molgenis.data.meta.DefaultPackage.PACKAGE_DEFAULT;
-import static org.molgenis.security.acl.SidUtils.createUserSid;
+import static org.molgenis.data.security.SidUtils.createUserSid;
 import static org.molgenis.security.core.runas.RunAsSystemAspect.runAsSystem;
 import static org.molgenis.security.core.utils.SecurityUtils.getCurrentUsername;
 import static org.testng.Assert.assertEquals;
