+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/importservice/VcfImportServiceIT.java
@@ -9,7 +9,7 @@ import org.molgenis.data.security.EntityTypeIdentity;
 import org.molgenis.data.security.PackageIdentity;
 import org.molgenis.data.security.auth.User;
 import org.molgenis.data.vcf.model.VcfAttributes;
-import org.molgenis.security.PermissionService;
+import org.molgenis.security.core.PermissionService;
 import org.molgenis.security.core.PermissionSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.acls.model.ObjectIdentity;
@@ -25,7 +25,7 @@ import static java.util.Objects.requireNonNull;
 import static org.molgenis.data.DatabaseAction.ADD;
 import static org.molgenis.data.meta.DefaultPackage.PACKAGE_DEFAULT;
 import static org.molgenis.data.meta.UploadPackage.UPLOAD;
-import static org.molgenis.security.acl.SidUtils.createUserSid;
+import static org.molgenis.data.security.SidUtils.createUserSid;
 import static org.molgenis.security.core.utils.SecurityUtils.getCurrentUsername;
 
 public class VcfImportServiceIT extends ImportServiceIT
