+++ molgenis-security-core/src/main/java/org/molgenis/security/core/PermissionServiceImpl.java
@@ -1,8 +1,5 @@
-package org.molgenis.security;
+package org.molgenis.security.core;
 
-import org.molgenis.data.security.auth.Role;
-import org.molgenis.data.security.auth.User;
-import org.molgenis.security.core.PermissionSet;
 import org.springframework.security.acls.model.MutableAcl;
 import org.springframework.security.acls.model.MutableAclService;
 import org.springframework.security.acls.model.ObjectIdentity;
@@ -12,8 +9,6 @@ import org.springframework.stereotype.Component;
 import java.util.Map;
 
 import static java.util.Objects.requireNonNull;
-import static org.molgenis.security.acl.SidUtils.createRoleSid;
-import static org.molgenis.security.acl.SidUtils.createUserSid;
 
 @Component
 public class PermissionServiceImpl implements PermissionService
@@ -23,30 +18,6 @@ public class PermissionServiceImpl implements PermissionService
 	public PermissionServiceImpl(MutableAclService mutableAclService)
 	{
 		this.mutableAclService = requireNonNull(mutableAclService);
-	}
-
-	@Override
-	public void grant(Map<ObjectIdentity, PermissionSet> objectPermissionMap, Role role)
-	{
-		grant(objectPermissionMap, createRoleSid(role));
-	}
-
-	@Override
-	public void grant(ObjectIdentity objectIdentity, PermissionSet permissionSet, Role role)
-	{
-		grant(objectIdentity, permissionSet, createRoleSid(role));
-	}
-
-	@Override
-	public void grant(Map<ObjectIdentity, PermissionSet> objectPermissionMap, User user)
-	{
-		grant(objectPermissionMap, createUserSid(user));
-	}
-
-	@Override
-	public void grant(ObjectIdentity objectIdentity, PermissionSet permissionSet, User user)
-	{
-		grant(objectIdentity, permissionSet, createUserSid(user));
 	}
 
 	@Override
